reorganize codes and files, add references
