+++ examples/transfer/plot_comparison_pipelines.py
@@ -10,12 +10,10 @@ Gaussian distribution.
 import numpy as np
 import matplotlib.pyplot as plt
 
-from pyriemann.datasets import make_gaussian_blobs
 from pyriemann.classification import MDM
-from pyriemann.utils.base import sqrtm, powm, expm
+from pyriemann.datasets.simulated import make_example_transfer_learning
 
 from sklearn.pipeline import make_pipeline
-from sklearn.utils import check_random_state
 from tqdm import tqdm
 
 from pyriemann.transfer import (
@@ -31,88 +29,13 @@ from pyriemann.transfer import (
 )
 
 
-def make_example_transfer_learning(N, class_sep=3.0, class_disp=1.0,
-                                   domain_sep=5.0, theta=0.0,
-                                   random_state=None):
-    """Generate source and target toy datasets for transfer learning examples
-
-    N : how many matrices to sample on each class for each domain
-    class_sep : how separable the classes from each domain should be
-    class_disp : intra class dispersion
-    domain_sep : distance between global means for each domain
-    theta : angle for rotation
-    random_state : int, RandomState instance or None, default=None
-    Pass an int for reproducible output across multiple function calls.
-    """
-
-    rs = check_random_state(random_state)
-    seeds = rs.randint(100, size=4)  # one seed for each sample_gaussian
-
-    # the examples considered here are always for 2x2 matrices
-    n_dim = 2
-
-    # create a source dataset with two classes and global mean at identity
-    X_source, y_source, means_source = make_gaussian_blobs(
-        n_matrices=N,
-        n_dim=2,
-        class_sep=class_sep,
-        class_disp=class_disp,
-        input_centers=None,
-        return_centers=True,
-        center_dataset=True,
-        random_state=seeds[0],
-        n_jobs=1)
-
-    # create target dataset based on the source dataset
-    X_target, y_target = make_gaussian_blobs(
-        n_matrices=N,
-        n_dim=2,
-        class_sep=class_sep,
-        class_disp=class_disp,
-        input_centers=means_source,
-        return_centers=False,
-        center_dataset=True,
-        random_state=seeds[1],
-        n_jobs=1)
-
-    # move the points in X_target with a random matrix A = P * Q
-
-    # create SPD matrix for the translation between domains
-    Pv = rs.randn(n_dim, n_dim)  # create random tangent vector
-    Pv = (Pv + Pv.T)/2  # symmetrize
-    Pv /= np.linalg.norm(Pv)  # normalize
-    P = expm(Pv)  # take it to the manifold
-    P = powm(P, alpha=domain_sep)  # control distance to identity
-    P = sqrtm(P)  # transport matrix
-
-    # create orthogonal matrix for the rotation part
-    Q = np.array([[np.cos(theta), -np.sin(theta)],
-                  [np.sin(theta), np.cos(theta)]])
-
-    # transform the data points from the target domain
-    A = P @ Q
-    X_target = A @ X_target @ A.T
-
-    # create array specifying the domain for each epoch
-    domains = np.array(
-        len(X_source)*['source_domain'] + len(X_target)*['target_domain']
-    )
-
-    # encode the labels and domains together
-    X = np.concatenate([X_source, X_target])
-    y = np.concatenate([y_source, y_target])
-    X_enc, y_enc = encode_domains(X, y, domains)
-
-    return X_enc, y_enc
-
-
 # choose seed for reproducible results
-seed = 90
+seed = 100
 
 # we consider several types of pipeline for transfer learning
 # dct : no transformation of dataset between the domains
 # rct : re-center the data points from each domain to the Identity
-# rpa : re-center, stretch and rotate
+# rpa : re-center, stretch and rotate (Riemannian Procrustes Analysis)
 # tlmdm : ???
 # calibration : use only data from target-train partition for the classifier
 scores = {meth: [] for meth in ['dummy', 'rct', 'rpa', 'calibration']}
@@ -121,10 +44,10 @@ scores = {meth: [] for meth in ['dummy', 'rct', 'rpa', 'calibration']}
 # are generated by the same generative procedure with the SPD Gaussian
 # and one of them is transformed by a matrix A, i.e. X <- A @ X @ A.T
 X_enc, y_enc = make_example_transfer_learning(N=100,
-                                              class_sep=0.50,
+                                              class_sep=0.75,
                                               class_disp=1.0,
                                               domain_sep=5.0,
-                                              theta=np.pi/4,
+                                              theta=3*np.pi/5,
                                               random_state=seed)
 
 # object for splitting the datasets into training and validation partitions
@@ -178,6 +101,9 @@ for target_train_frac in tqdm(target_train_frac_array):
             pipeline.score(X_enc_test, y_enc_test))
 
         # (2) rct pipeline: recenter the data from each domain to identity
+        # see Zanini et al. "Transfer Learning: A Riemannian Geometry Framework
+        # With Applications to Brain–Computer Interfaces". IEEE Transactions on
+        # Biomedical Engineering, vol. 65, no. 5, pp. 1107-1116, August, 2017
 
         # instantiate the pipeline
         step1 = TLCenter(target_domain=target_domain)
@@ -203,6 +129,9 @@ for target_train_frac in tqdm(target_train_frac_array):
         scores_cv['rct'].append(pipeline.score(X_enc_test, y_enc_test))
 
         # (3) rpa pipeline: recenter, stretch, and rotate
+        # see PLC Rodrigues et al “Riemannian Procrustes analysis: transfer
+        # learning for brain-computer interfaces”. IEEE Transactions on
+        # Biomedical Engineering, vol. 66, no. 8, pp. 2390-2401, December, 2018
 
         # instantiate the pipeline
         step1 = TLCenter(target_domain=target_domain)
@@ -212,7 +141,7 @@ for target_train_frac in tqdm(target_train_frac_array):
             centered_data=True)
         step3 = TLRotate(
             target_domain=target_domain,
-            metric='riemann')
+            metric='euclid')
         clf = TLClassifier(
             target_domain=target_domain,
             clf=MDM())
@@ -272,9 +201,9 @@ for meth in scores.keys():
         scores[meth],
         label=meth,
         lw=3.0)
-ax.legend(loc='center right')
-ax.set_ylim(0.45, 0.85)
-ax.set_yticks([0.5, 0.6, 0.7, 0.8])
+ax.legend(loc='upper right')
+ax.set_ylim(0.45, 0.75)
+ax.set_yticks([0.5, 0.6, 0.7])
 ax.set_xlim(0.00, 0.21)
 ax.set_xticks([0.01, 0.05, 0.10, 0.15, 0.20])
 ax.set_xticklabels([1, 5, 10, 15, 20])
