+++ pyriemann/transfer/__init__.py
@@ -1,0 +1,20 @@
+from .methods import encode_domains, decode_domains, TLSplitter
+
+from .pipelines import (
+    TLDummy,
+    TLCenter,
+    TLStretch,
+    TLRotate,
+    TLClassifier,
+    TLMDM,
+)
+
+__all__ = ["encode_domains",
+           "decode_domains",
+           "TLDummy",
+           "TLSplitter",
+           "TLCenter",
+           "TLStretch",
+           "TLRotate",
+           "TLClassifier",
+           "TLMDM"]
